add constraint to csearch workflow
