+++ stjames/__init__.py
@@ -19,3 +19,4 @@ from .correction import *
 from .solvent import *
 from .mode import *
 from .status import *
+from .constraint import *
