+++ stjames/opt_settings.py
@@ -1,21 +1,7 @@
 from pydantic import Field, PositiveFloat, PositiveInt
 
-from .base import Base, LowercaseStrEnum
-
-
-class ConstraintType(LowercaseStrEnum):
-    """Different sorts of constraints."""
-
-    BOND = "bond"
-    ANGLE = "angle"
-    DIHEDRAL = "dihedral"
-
-
-class Constraint(Base):
-    """Represents a single constraint."""
-
-    constraint_type: ConstraintType
-    atoms: list[int]  # 1-indexed
+from .base import Base
+from .constraint import Constraint
 
 
 class OptimizationSettings(Base):
